Refactored for v0.2 release
Completely removed old system of individual functions
in favor of classes to do the work. The original functions
had a fatal flaw when working with complex transactions.

Added new system that implement the Calculator interface and
added two calculators BCMath and Float. Added documentation
for the new system as well.

Signed-off-by: RJ Garcia <rj@bighead.net>